backend/requirements.txt
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
certifi==2020.12.5
cffi==1.15.0
channels==3.0.3
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
decorator==5.0.7
Django==3.2
django-admin-rangefilter==0.8.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.9.0
pyOpenSSL==21.0.0
pytz==2021.1
requests==2.25.1
service-identity==21.1.0
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==4.0.0
urllib3==1.26.4
wcwidth==0.2.5
zope.interface==5.4.0
